# api-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.createDeviceRequest:
    properties:
      name:
        type: string
    type: object
...
  handlers.deviceResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      subscription_url:
        type: string
      wg_address:
        type: string
      wg_public_key:
        type: string
      xray_uuid:
        type: string
    type: object
...
  handlers.userNodeItem:
    properties:
      country:
        type: string
      name:
        type: string
      region:
        type: string
      status:
        type: string
    type: object
  handlers.userPlanFeature:
    properties:
      included:
        type: boolean
      text:
        type: string
    type: object
  handlers.userPlanItem:
    properties:
      duration_days:
        type: integer
      features:
        items:
          $ref: '#/definitions/handlers.userPlanFeature'
        type: array
      id:
        type: string
      max_devices:
        type: integer
      name:
        type: string
      prices:
        items:
          $ref: '#/definitions/handlers.userPlanPrice'
        type: array
      purchasable:
        description: |-
          Derived the same way as the admin side: a plan with no priced duration
          cannot be ordered, independent of is_active (which only gates whether
          it is shown or requestable at all).
        type: boolean
      speed_limit:
        type: integer
      traffic_limit:
        type: integer
    type: object
  handlers.userPlanPrice:
    properties:
      duration_days:
        type: integer
      price_cents:
        type: integer
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: VPN Panel Management API
  title: Proxima VPN API
  version: 0.1.0
paths:
...
  /user/devices:
    get:
      description: Returns all devices for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.deviceResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List devices
      tags:
      - user-portal
    post:
      consumes:
      - application/json
      description: Registers a new device for the authenticated user
      parameters:
      - description: Device name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.deviceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create device
      tags:
      - user-portal
  /user/devices/{id}:
    delete:
      description: Removes a device belonging to the authenticated user
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete device
      tags:
      - user-portal
  /user/nodes:
    get:
      description: Returns name and location of the nodes the user's plan can use
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.userNodeItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List my available nodes
      tags:
      - user-plans
...
  /user/plans:
    get:
      description: Returns all active subscription plans
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.userPlanItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List available plans
      tags:
      - user-plans
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"